+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.21 branch
 ### 2.21.3
+* fixed templates for `BootstrapFormHelper` with input groups;
 * fixed little bug for `BootstrapFormHelper::submit()` method.
 
 ### 2.21.2